Update i3 config to disable scroll wheel actions and enhance backpages script for process management

+++ lib/backpages.sh
@@ -2,8 +2,12 @@
 # 查找 backpages/main.py 的进程
 PIDS=$(pgrep -f "/data/codes/ranger/backpages/main.py")
 
-if [ -n "$PIDS" ]; then
-  # 如果发现已有进程，则杀掉所有相关进程
-  kill -9 $PIDS
+if [ "$1" == "show" ]; then
+    ~/miniconda3/envs/usr/bin/python /data/codes/ranger/backpages/main.py --show
+else
+  if [ -n "$PIDS" ]; then
+    # 如果发现已有进程，则杀掉所有相关进程
+    kill -9 $PIDS
+  fi
+  ~/miniconda3/envs/usr/bin/python /data/codes/ranger/backpages/main.py &
 fi
-~/miniconda3/envs/usr/bin/python /data/codes/ranger/backpages/main.py &
